Move CrawledUriService and DAO to crawleduri.py
